GfxBox added !
Now we can create either sprites (using an image texture) or simple rectangle boxes with plain colors
The only difference between a Sprite and a Box is the textureID and the plain color respectively.

A box can be, like any sprite, autorotating, having a filter color (that is not really useful as this is a plain color so we can give the final requested color to it), can be blinking, ...

BUT the code inside the shader is not very clean
I should organize it a little bit more, creating functions and so on ^^

---------------------------------
Next step will be to add an OVAL gfx element : the main part of the modifications, will be inside the fragment shader in order to know if we are inside the oval part or not : it will be like any ShaderToy.com example ;)

I thought about a feature displaying a border around the rectangle but it could be part of the common gfx code : a sprite could also have a border around it
So may be the border highlight will be in the Gfx class instead just being in the GfxBox class. It could so be used for the OVAL class too

to be continued ....